Started MySQL database integration for test and production servers, adding database.<dev|test>.<hst|jwst>.py files in configs, modified install, working for SQLITE.   The database.test files are untested at this point.
Changed import_files.py to make missing files a warning rather than an error so that all files will be known to CRDS and won't prevent hst.pmap from certifying.   Nominally this applies to HST where there are files in rmaps which don't exist in /grp/hst/cdbs.

Added a warning and defined instrument and filekind as "UNKOWN" when a submitted file has a problem getting file properties for some reason,  generally due to missing files in /grp/hst/cdbs.

Switched from "NOT PRESENT" to "UNDEFINED" for missing header keywords.



git-svn-id: https://aeon.stsci.edu/ssb/svn/crds_server/trunk@469 37088bc5-bd3f-4c5b-b9de-119bb1e8f7e3