Added more progress and size info to submission status logging.
Changed submission copy/link processing to link-only to save inline copy of
gigabytes of data; must perform some secondary action like background ownership
assumption since linked files will retain the original owner.  The original
owner will not be the CRDS server for files submitted by direct copy to the
ingest directory which is the current norm for the command line interface.  Nominal
CRDS approach is to re-own files using a cron job by making a duplicate copy and
deleting the original.   This is a case where chown is disallowed even though the 
server owns the controlling directory and can both read and delete the files,  so
it appears to be a compromise to simplicity rather than making sense.   Chowning
is subtle,  it's not even possible to give away files to prevent "quota bombing".
This may not be a permanent solution and is not required when files are uploaded 
through the server by HTTP since then crds automatically owns the files.

Deleted large dead comment string.


git-svn-id: https://aeon.stsci.edu/ssb/svn/crds_server/trunk@2479 37088bc5-bd3f-4c5b-b9de-119bb1e8f7e3