Updated files state definitions and certification process to ensure that files that are not yet fully confirmed don't pass the "reference must exist in CRDS" rmap validation check.
This addresses a recent operational failure where a file submitter used an rmap
from an aborted but uncancelled delivery as the basis for a manual rmap update.
Because the failed delivery was still sitting around in an unconfirmed state,
the new reference files the generated rmap referenced were still in CRDS as
models in the "uploaded" state.    When the rmap was erroneously used as the basis
for a manual update,  it continued to reference unconfirmed files but was not
rejected.   When the unconfirmed references were later cancelled,  after the derived
rmap was already confirmed,  the derived rmap referred to files not actually in CRDS.
This update is intended to cause rmap's that reference unconfirmed files to be
rejected during certification.