Removed the update_server_cache subprocess calls from the Set Context and Mark Files Bad web functions after observing how they interact with unit tests.
Run as a subprocess, update_server_cache is inheriting the full-up server
environment rather than the test environment.  Since the unit tests move the
server to an offline URL, the cache sync operations fail and wind up doing 10
minutes of retries for the Set Context and Mark Files Bad unit tests.   This
is a huge waste of time,  but otherwise OK.

A worse prospect would be a situation where the update_server_cache script
points to the server under test correctly but is misconfigured to use the real
cache instead of the test cache.  This would cause update_server_cache to write
the test context into the real cache, which if shared with the pipeline, would
cause the pipeline to almost certainly start using the wrong context.    The
worst case scenario would be if the context "worked" in the pipeline,  escaping
notice.

This change restores the classic behavior of the server, which is to sync the
server cache's config area (which records context) only by cron.  During unit
tests, cron is turned off, so it is not an issue.  During normal operations,
after the operator does Set Context, and syncs the pipeline cache, the server
may lag by 5 minutes or so updating the default context it uses for services,
but no more.   Ditto of Mark Files Bad.

Since we're not yet using server+pipeline shared CRDS caches, backing off some
seems like the prudent thing to do,  both to avoid potential cache errors,
and to save significant amounts of time running unit tests.   The complexity
of it argues against fixing it any further.